requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pyyaml>=6.0
numpy>=1.24.0
pydantic>=2.0.0
